feat: add basic "Tools" menu
This commit introduces a basic "Tools" menu with two placeholder items: "Tool 1" and "Tool 2". These tools currently only log a message to the console when clicked.

This is a preliminary step towards implementing more advanced features in the future. The "TODO" comments indicate where additional menu items and functionality will be added in subsequent commits.